Исправил тип данных PayablePlan

+++ ValidationRules.Storage/Model/Erm/Bill.cs
@@ -6,7 +6,7 @@ namespace NuClear.ValidationRules.Storage.Model.Erm
     {
         public long Id { get; set; }
         public long OrderId { get; set; }
-        public long PayablePlan { get; set; }
+        public decimal PayablePlan { get; set; }
         public DateTime BeginDistributionDate { get; set; }
         public DateTime EndDistributionDate { get; set; }
         public int BillType { get; set; }
